CLOUD-4062: Refactor system service
Fixed admin to work with the new way for setting the system in the settings component.
Reorganized markup for settings component.
Light refactor of settings component ts.
Major refactor of the system service.
Moved most of the users functions and data into new class UserManager.
Added watcher for tracking the number of subscribers for the internal poll.

Branch: cloud_19.3