X/email object
This is an email object that intends to centralize the existing test related email related functionality. 

First thing to note is that the get_random_email function is a static method so you don't need to instantiate the object:>>> from email_access import Email
>>> Email.get_random_email(sendemail=True)
'noptixautoqa+sendemail1692898245.40853878@gmail.com'

I have also chosen to have each function either to login separately or at least check to see if there is an existing session. This is inefficient, but given this is used in tests, I think trading the efficiency for flexibility is a win. 

I have tested each function but I haven't tested how it integrates with existing code.