Adding clientProtocol/uriProtocol to webadmin
## Problem

We need access to the `uriProtocol` (also called `clientProtocol`) in WebAdmin now, because we are opening links directly to the desktop client

## Solution

Modify the `apply_customizations.py` script to allow the `uriProtocol` value from `description.json` to be preserved. Then update existing logic in WebAdmin to take that new value from Description.

In the end this gives us the `nx-nvm` value that we use when opening the desktop client (nx-vms://cloud-test.com). This value changes for our partners so it has to come through these customization files.