CLOUD-11273: Add missing security level parameter
In `front_end/apps/setup-wizard/src/app/services/wizard-state.service.ts`, there is a call to the Rest API v2's `setupLocalSystem()` method passing in `securityLevel` as a parameter. However, the `setupLocalSystem()` method in our Rest API v2 service is currently missing this `securityLevel` parameter. There was a FIXME comment in the Wizard State service calling out that the parameter should exist in the v2 API method, but was missing. This MR adds in the missing `securityLevel` parameter (based on the way the parameter is used in the Rest API v1 service), and removes the FIXME comment.