conversion of activate.robot
Conversion of activate.robot to python.

Modifies the email_access object so that emails that are scanned for activation links are marked as read.

this was failing:
    api = CloudPortalAPI()
    user_data = api.get_account_data(random_email, rb.BASE_PASSWORD)

CloudSession.py was modified to update the header on logout to fix that. 

There are a few tests that don't work because of outstanding bugs.