CLOUD-11302, CLOUD-11303, CLOUD-11304, CLOUD-11305: Channel Partner Updates
**Changes**
- Removes store usage from `create-system-group.component`
- Removes `NxSystemGroupsService` from root and moved usage strictly into `organization.component.ts`
- Show actions bar in `groups-cards.component.html` when no groups/systems are available (primarily for add group button)
- Converts some observables into signals

**Bug fixes**
<details><summary>Fixes Organization pages not working</summary>
Changes to the `customAccountPropertFactory` caused `NG0203` error due to not being called in component constructor, moving the call of the `customAccountPropertyFactory` to component constructor fixed the issue
</details>
<details><summary>Fixes System pages not working (shared & personal)</summary>
Systems were changed to grab from the store at one point. The CP websocket is in charge of populating this data, so when moving the `groups-service` connection strictly to `organization.component.ts` from `home.component.ts` data would not be grabbed from the websocket. Fixed this by grabbing from the `NxSystemsService`
</details>
<details><summary>Console errors reading undefined</summary>
`role` would sometimes not be defined in some services so adding optional chaining cleared those errors. Although this issue could go deeper, decided to keep the scope of the MR closer to CP
</details>