Change token to temporaryUserToken
## Problem

For temporary users feature we pass a token as a query param. We updated the spec for the name of this temporary token from `token` to `temporaryUserToken`

## Solution

Update our logic so instead of parsing the `token` for a specific token string, we can just look at the query param name. Use this opportunity to clean up a couple of the variable names as well