CLOUD-11309 - Adding typing to motion mask
## Problem

We want to add typing to `MotionMaskRenderer.ts` and `MotionMaskState.ts`

## Solution

There is nothing too complicated here. A lot of it was already typed, but it just took going through and adding the correct type in a few missing cases. I created an `ActionType` enum instead of just using string comparisons in one place.