CLOUD-11338: Added offline system support the permission manager
# Problem
The new permission manager only worked with online systems. An offline system can many things to be undefined causing error in maybe places.

Users in particular was problematic and had an infinite loop that froze the app.

# Solution
Add more error handling and rely on clouddb for user info if the system couldn't be reached.

# Changes
- Added error handling where things could return empty. Most cases it was related to groups.
- Added a check so that only online systems attempted to use cookie login.
- Updated the system poll so that it fetched cloud users if the system was offline.
- Added clouddb support to the permission manager.