CLOUD-11280: Add loading state to language dropdown
The language picker in the Account Settings page updates immediately upon selection of a new language, without having to click the save button. However, there is a slight delay between when the language is selected and when the new language is applied to the page, and there is no loading UI associated with this delay. This can lead to some confusion, and resulted in the bug report in the associated ticket.

This MR adds a loading state to the language picker UI. The html/CSS for the input loading state was ported from:
https://gitlab.nxvms.dev/dev/cloud_portal/-/blob/develop/front_end/libs/components/process-button/process-button.component.html#L5-13
https://gitlab.nxvms.dev/dev/cloud_portal/-/blob/develop/front_end/libs/components/process-button/process-button.component.scss#L55-109

![Screen_Recording_2023-09-08_at_11.43.22_AM](/uploads/185cf1b6a34da2eac31e0df5e99b942c/Screen_Recording_2023-09-08_at_11.43.22_AM.mov)