CLOUD-11284: Fix tooltip position for Add User dialog and fix tooltip arrow color
# Problem
The tooltip position was fixed for the User Permissions dropdown, but it wasn't fixed for the Add User dialog's permissions. The tooltip arrow color was also wrong as it was using the alternate tooltip's color

# Solution
Add directives to this dropdown to fix the positioning. Fix CSS to display proper color for alternate tooltips