CLOUD-10793: Fix 2FA on session timeout in WebAdmin
This MR fixes a bug:  
Previously, when logged in to WebAdmin as a cloud user with 2FA enabled, and trying to disconnect the system from cloud after the session timed out, a cloud oauth page would open in a new tab for the user to re-authenticate. After entering password in the cloud oauth page, the user would be redirected back to the WebAdmin disconnect system page, which would be stuck in a loading state. The oauth re-authentication flow did not ask for a 2FA verification code after entering password, and the user would not be able to successfully refresh their session and disconnect the system from cloud.

This bug was occurring because WebAdmin was not tracking whether the cloud user had 2FA enabled for the system, and subsequently was also not handling 2FA in the re-authentication flow for cloud users on session timeout. This MR makes the following changes to address this:

1. On oauth login for a cloud user in WebAdmin, there is an existing check to see whether 2FA is required for the system, which is why 2FA authentication is working on login. This MR saves that result as a boolean `system2faEnabled` flag in local storage, so that WebAdmin has a way of knowing whether 2FA is required for re-authentication on session timeout. On session timeout, if 2FA is required for the system based on the local storage flag, a `renew2FA` state is passed to the updateSession() handler and then to the cloud authorization page, otherwise a `renewWeb` state is passed. The cloud authorization page handles 2FA based on this passed in state.

1. After session timeout in WebAdmin and successful login in the cloud authorization page, the auth page was storing the oauth code in local storage and then closing the window. This is intended behavior for an auth request coming from Nx Cloud, where the client app is on the same domain, and so the code in local storage would be accessible by client app. However, WebAdmin is on a different domain than cloud auth, and so for this flow what we want is for the auth page to redirect back to WebAdmin with the oauth code passed as a URL parameter. This MR adds an additional check to confirm whether the oauth client is `cloud` before stepping into the local storage save/close window logic; if not it proceeds to the logic of redirecting back to the redirect link with the code passed as a URL param. This MR also sets the redirect link if the client type is `renewSessionWeb2FA` (previously `this.redirectLink` was undefined at the redirect call).


Successful 2FA and disconnecting of system from cloud after session timeout in WebAdmin (note - at the end, after disconnecting a system from cloud when logged in as a cloud user, the user will be logged out):
![Screen_Recording_2023-09-18_at_10.00.20_AM](/uploads/9d56910ab500a508c1510c0c1b84c856/Screen_Recording_2023-09-18_at_10.00.20_AM.mov)