CLOUD-11347 - Temp user icon in header
## Problem

For temp users we need to update the icon in the header

## Solution

This data exists on the user, but it wasn't being added to the `Account` type in the `newLocalAccount` function. I had to update some types and pass type through. Then the logic is simple just check if type is tempLocal

| BEFORE | AFTER |
| ------ | ------ |
|    ![Screenshot_2023-09-13_at_3.41.16_PM](/uploads/131b3a5b2ac0a7a055211edeac11e12f/Screenshot_2023-09-13_at_3.41.16_PM.png)    |    ![Screenshot_2023-09-13_at_3.40.25_PM](/uploads/1d49035d12a62398a6bb19c92f87f401/Screenshot_2023-09-13_at_3.40.25_PM.png)    |