Small timing changes to fix motion mask
## Problem

Currently the motion zone is completely broken. This is due to an in progress refactor that is on pause as other work continues

## Solution

There were some timing issues that were only causing visible problems with the Motion Mask stuff, because it was sort of reliant on timings of components and data being setup. Fixed this by moving some of my refactored code around so it is called in the proper order. Added a first render variable as a temporary fix for motion mask canvas stuff.

In the end this isn't great but it is still in progress of refactor. This fixes the existing bugs but doesn't undo some of the refactor work I was in the middle of.