CLOUD-11280: Fix language dropdown design bugs
This MR fixes a couple design bugs in the language dropdown that were introduced by [#4519](https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/4519). #4519 added a loading state to the language dropdown in the Nx Cloud `Account Settings` page by wrapping the dropdown button in div. The div renders if `isLoading$$()` is `false`, otherwise an `<ng-template>` of the button in a loading state is rendered:
![Screen_Recording_2023-09-20_at_9.42.32_AM](/uploads/c8c7339f27a07598b518109ffc5e17f7/Screen_Recording_2023-09-20_at_9.42.32_AM.mov)

This worked fine for the language dropdown in the Cloud `Account Settings` page, but several different versions of the language dropdown in both WebAdmin and Nx Cloud are all created using this same base component/html template, and the html template changes in #4519 created design bugs in some of those other versions of the language dropdown:

1. The background color of the header language dropdown changed from transparent to light grey. This is because there is a declaration on L11 of  `language-webadmin.component.scss` making the background transparent, and the selector for this declaration 
```
.dropdown {
  & > button#dropdownMenuButton {
    background : transparent !important;
  }
}
```
required the element to be a direct child of `.dropdown`. Since #4519 wrapped `button#dropdownMenuButton` in a div to conditionally render it when `isLoading$$()` is `false`, it broke this direct child relationship and therefore this css declaration was no longer being applied. Previously, there were no `button#dropdownMenuButton` elements that were not a direct descendant of `.dropdown`, so it seems that part of the css selector is no longer necessary, so this MR just removes it to fix this issue.

2. The loading button added in #4519 was intended just for the version of the language dropdown that appears on the Cloud Account Settings page. However, the wrapper div that conditionally renders the loading button was wrapped around all versions of the dropdown button, and so all of them showed a loading state that did not match the design of those dropdown buttons. The different versions of the dropdown button are rendered based on `inHeader`, `inFooter`, `newHeader`, and `short` flags. This MR moves the loading wrapper and updates the conditional rendering logic so that it only appears for the version of the language dropdown in the account settings page - a long (not short) button in the page body (not header or footer).
  
<br />

***

<br />

Before: 
![Screen_Recording_2023-09-20_at_9.43.24_AM](/uploads/7722ef9cd7711ae456d2c2a412eac9a5/Screen_Recording_2023-09-20_at_9.43.24_AM.mov)

After:
![Screen_Recording_2023-09-20_at_9.39.32_AM](/uploads/6c81263c4754beb5e2a42686cf9e4eeb/Screen_Recording_2023-09-20_at_9.39.32_AM.mov)