CLOUD-10656 -refactor of getLink method
# Problem
"Install Nx Witness to view your System" window appears if client is already installed.

# Solution
Added a method that checks to see if it's a possible to exchange a code for a token. If it succeeds, that means that the client is not installed, so the "Install Nx Witness" dialog is shown; if it fails, we can assume that the client is installed, so the dialog isn't shown and the desktop client is opened.