CLOUD-11515: Updated the check for setting current user
# Problem
user.id only works for local users and the check doesn't work on cloud.

# Solution
Bring back email check which was working, but adjust it to account for user types. Local users don't have an email and in webadmin mode currentUserEmail can be ''. In this case all local users without email will be set to the current user.