CLOUD-11447 - checkbox show state even if disabled
## Problem

In the camera settings there is a checkbox for `Audio Enabled`. There are more underlying issues for this ticket, but for Web's side we can clear up some confusion. We currently don't allow a checkbox to be checked/true if it is disabled

## Solution

Remove some logic that was forcing checkbox to always be false if it is disabled.

## Note

There is a chance some components were relying on this behavior. I tried to play around with some other checkboxes but didn't see any immediate issues. Will be something to keep an eye on though

| **BEFORE** | **AFTER** |
| ------ | ------ |
|    ![image](/uploads/eda96c4b1600f8893bad43d89cca30ed/image.png)    |    ![image](/uploads/23af07a50f435e384685198c7831629f/image.png)    |