CLOUD-11365, CLOUD-11440, CLOUD-11449, CLOUD-11452, CLOUD-11510, CLOUD-11511, CLOUD-11513: Refactoring system user page
# Problem
The apply service doesn't play well with the multi-select-dropdown which wreaks havoc with form change detection

# Solution
Rip and replace the apply service and watchers in favor of reactive forms which is the more angular way of dealing with forms.
In the process of doing this also created extensions of FormGroup (NxFormGroup) and FormControl (NxFormControl) to help with future rip and replaces of the old apply service + watchers.

# Changes
- Replaced NgForm with NxFormGroup. This allows us to build a form and let angular check for changes.
- Migrated filteredGroups to a computed signal so that it can auto update. A side benefit is that groups also has a signal for future improvements throughout the app.
- Cleared up the edit user process so that they are now correct.
- Refactored the apply component so that it can be used independently of the original apply service. This will help with killing off the apply service in the future.
- Created a newer version apply service that's striped down and is mainly used for navigating.

# Todo
- Add support to the apply-v2-service so that it can support multiple form groups on a page. (Maybe not necessary with how reactive forms work)