CLOUD-11405: Bugfix - Edit name of local user
# Problem
An owner could not edit the name of a local user, both in WebAdmin and Cloud. After changing a local user's name and clicking save, the change would not update the backend/would not persist after a page reload.  
![video_1280](/uploads/67d96304aec9f86879282680af1598c2/video_1280.mp4)

# Solution
After an owner makes changes to a user and clicks save, the frontend would call the `modifyUser()` method of the `UserWithGroupsManager` service, passing in the updated user object. `modifyUser()` would validate/clean up the updated user object before calling a patch method on the V3 user endpoint. Within `modifyUser()`, the `fullName` property of the user object was being deleted before passing the user object to the API. This was because in the past, the API had issues when trying to update the name and groups of a user simultaneously. However, this API issue seems to be resolved as of `6.0.0.37561`, so this MR removes the code that deletes `user.fullName` and leaves a comment explaining the previous API issues. 

Updating a local user's name is fixed in both WebAdmin and Cloud with this change:  
![2_Screen_Recording_2023-10-04_at_10.55.26_AM](/uploads/38e61dc313413c1cd32fbd4e2c347791/2_Screen_Recording_2023-10-04_at_10.55.26_AM.mov)