CLOUD-11557 - Camera settings new dropdown
## Problem

We have a new dropdown that will allow us to reduce complexity in our components. This is the first MR to implement it.

## Solution

Remove the old dropdown. Change the stored values in the component. This will be more evident why this is an improvement once we remove the need for watchers.

<div align="center">

### There should be no visible changes
</div>

![image](/uploads/5c8a833d9ce37e8053fa26ec7e53f70e/image.png)