CLOUD-11554 - camera settings bug if user has edit permissions
## Problem

If a user has View permissions but not edit permissions for a camera, it can still show up in the edit camera settings. This can happen if the user has edit permissions for a different camera. Permission table below

| | Camera1 | Camera2 |
| ------ | ------ | ------ |
|   View     |    ✅    |    ✅    |
|   Edit     |    🚫    |    ✅    |

## Solution

We need to do a check on each camera for the edit permissions. We can do this in the parse camera stage and add a parameter to the camera object. Then in the 3 places necessary we can check the `canEdit` value on the camera.