CLOUD-11448 - logout temp user when session expires
## Problem

Temporary Users have expiration timers on their sessions. When this timer is up they are no longer allowed to be logged in. Instead of prompting them to relogin we should just instantly log them out.

## Solution

I tried to use the existing session interceptor but there was no crossover so it made sense to make a new one. I also tried to use the `AccountService.logout` function, but you can't inject `AccountService` in the interceptor because things aren't loaded properly. That was a lot of work to fix (I tried..) so I went with the explicit straight forward method.