CLOUD-11344: Add cloud owner name in WebAdmin
## Problem

When a Power User views a cloud system in webadmin, only the cloud owner's email is shown and not their name. The intended behavior is to have both the cloud owner's name and email shown, which is what currently occurs in Cloud.

![image](/uploads/be266d75612b26c5e9844c576a4903d7/image.png){width=70%}

![image](/uploads/389cd86750e05bb7d18eacce379b3b75/image.png){width=70%}



In the admin settings component, the html template references `system.info.ownerFullName` to populate the system owner's name. In Cloud, the `ownerFullName` property exists under `system.info`, however in WebAdmin it is undefined. The reason for this difference is that system info is fetched differently in Cloud and WebAdmin. Cloud fetches system info from a `cdb/systems` endpoint which returns the cloud owner name for the system, whereas WebAdmin fetches system info from a `rest/v1/system/settings` endpoint, which doesn’t return the cloud owner name for the system.

## Solution

System info is fetched via the `getInfoAndPermissions()` method in `NxSystemOldModule`. For WebAdmin, the cloud owner's email is returned with system info, so we can use that to look up the owner from the `users` array, and get the owner's `fullName` there. 

Within the WebAdmin control flow of `getInfoAndPermissions()`, after fetching system info and users from the API, this MR adds one line to find the user whose email matches the cloud owner's email, and then sets `system.info.ownerFullName` as that user's `fullName`. With this change the cloud owner's name shows up in the WebAdmin UI. 

![image](/uploads/61a23f0f923b0fc57d9eca6f1e396e79/image.png){width=70%}