CLOUD-11588 - fixing bug where admins cant view cameras
## Problem

After fixing a previous but, we introduced a new bug where admins can't view camera settings

## Solution

Because Admins don't have specific permissions, but instead have global permissions, we can't just rely on the `canEditSpecificCamera` value.

In the simplest form we should check if a user is an admin or if they have specific permissions to edit a camera. This is the change we made here