Remove CloudPortalAPI.customization and its usage
It was only used in one method: CloudPortalAPI.bind_system(). This method wasn't used anywhere.

Also, when we need to bind a system we can still use CloudPortalAPI.connect() as we already do. This method does not require the user to specify the customization. It determines the customization on its own and calls cloud_api.System.bind in the end.