CLOUD-11598: Refactored logic for finding Power User Groups
# Problem
When a group is imported from LDAP it can form a cycle.
Ex: Group A -> Group B -> Group C -> Group B

# Solution
Created a new method that recursively checks the permission graph that ignores visited nodes when a cycle is detected. Visited nodes are ignored because by the time we run into them again we should have exhaustively looked through their list of parents. At any point if it was to return true the some method would end. While we recursively check parent nodes we add them into the powerUserGroups set reducing the depth we need to check on future iterations.

# Changes
- Added a set for tracking Power User groups
- Moved processing/checking power user groups from during processUser to processGroups. This will cut down on the work for checking power user inheritance. (n (users) -> once per poll interval)