CLOUD-11485: Hide the "Disconnect from My Account" button in WebAdmin
### Problem
In WebAdmin, the "Disconnect from My Account" button is visible if the logged in user is a cloud user. The intended behavior is for this button to not be visible in WebAdmin, and visible only in Cloud.

### Solution
Previously, the button was conditionally rendered when in a Cloud environment, or if the current user was a cloud user (`*ngIf="!environment.isLocal || user?.isCloud"`). This MR updates the *ngIf condition such that the button only renders in Cloud, and not when the current user is a Cloud user when in WebAdmin (`*ngIf="!environment.isLocal"`). With this change, the button is no longer visible in WebAdmin when logged in as a cloud user.
  
**Before:**  
![image-20231016152014756](/uploads/6b3aa8eb637f12ce4deeab0d66397c37/image-20231016152014756.png){width=50%}
  
**After:**  
![image-20231016151941557](/uploads/760149e7b4eb2f7924b9080134a1a3d3/image-20231016151941557.png){width=50%}