CLOUD-11668: Add barebones animations with Feature Flag
# What we're doing
We're slowly implementing Angular's reusable animations to the site. This push just adds 2 use cases for using animations, specifically on the `License Input` and the `Add User` dialog's "no-input-error" text. This will be added as a Feature Flag, but will be added to the CMS later

# Changes
Add basic reusable animations to `License Input` and the `Add User` dialog. Enable it via Feature Flags in the Repo, this will be added to the CMS in the future. We're still testing this and will add more features later