CLOUD-11322 - Login modal appears while processing OAuth token
## Problem

When we login using cloud on WebAdmin we grab the token and it's supposed to show a loading indicator. It's not showing the loading indicator

## Solution

What's going on is that the `login-webadmin.component.ts` is initialized multiple times so it loses context if it is currently processing an OAuth token. This multiple reload causes a bunch of issues and I've previously tried to fix it for other bugs but it's difficult. It's been baked in for so long and some things rely on the multiple reloads at this point.

This fix is just making a tiny service to store if we are currently processing an OAuth token. This bug isn't big enough to warrant a huge amount of changes