CLOUD-11556  - Adding Multi Select to the new Dropdown
## Problem

Previously we created a new dropdown component with basic functionality. We need the ability to have multi-select in the new dropdown

## Solution

* First pass was in https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/4755 , but refactored so that Single and Multi select dropdowns are different components. This isn't ideal, but it cleaned up the code tremendously so I think it's worth the tradeoffs.
* This MR adds the Multi-Select functionality to the dropdown. This has no effect on it visually. Only adding the functionality to support multi select
* It also creates a `NxMultiSelectDropdownItemComponent` to match what dropdown items in multi select dropdowns look like.
( **REMINDER** these functionalities are separate)

### NOTE
There are some limitations and workarounds needed for Typescript here. Because the type of `selected` changes based on the `Multiple` input value, we have to create types that fulfill this need. There are mostly Typescript solutions to this, but it's not perfect.

---

Multi Select **WITH** the new `NxMultiSelectDropdownItemComponent`

![image](/uploads/68c62096ea80923b64a7e90bb173d54e/image.png)

Multi Select with the basic dropdown item. This isn't useful in this state, but there is a high possibility MultiSelect will be needed with different looking dropdown items. This just shows the high flexibility of the dropdown

![image](/uploads/d00796090bc7c93af509641030dcd507/image.png)