CLOUD-11672 - View tab conditionally show edit camera settings
## Problem

In the view tab there is an "Edit Camera Settings" button that shows, but the logic for when it is rendered is wrong.

## Solution

We shouldn't be using `permission().editCamera` as that just tells us if there is any camera that could be edited. We need to know if a specific camera has edit permissions. This data lives on the `resourceAccessRights` on the user.

We already did this work but it was for the `NxCamera`. So we have to do it again for the `ViewCamera`. I don't know enough about the View tab or the `ViewCamera` and why it exists when we already have `NxCamera` to do a fancier refactor.

| Camera WITHOUT edit permissions | Camera WITH edit permissions |
| ------ | ------ |
|    ![image](/uploads/9e5225e83b16a2f0291f727cf63b881e/image.png)    |    ![image](/uploads/3fb2d631d03b77dbe5be7ca7af131cc7/image.png)    |