CLOUD-11597 - Bug with unauthorized cameras in camera settings
## Problem

If we load the camera and it is unauthorized AND ALSO it does not have a default aspect ratio set. Then we were internally setting the aspect ratio to `0` and the dropdown was not showing the "Auto" option.

## Solution

1. Change the camera parser to either set the aspect ratio to a real default aspect ratio, or to set it to null. This is clearer in using the data what is going on.
2. Remove the if check on when to display the "Auto" option in the dropdown since desktop client always shows this option.
3. Finally, there was a bug with the `DropdownItem` where the component wasn't initialized and it was trying to access a `ViewChild`.

All of these fixes are required to show the correct dropdown items

![image](/uploads/269cf642b9c34c59c258cf4271836ffe/image.png)