CLOUD-11589: Remove empty footer when viewing self in WebAdmin as a cloud user
# Problem

In WebAdmin, the user details page renders with an empty footer when logged in as a Cloud User and viewing myself. We don't want to show an empty footer in this case.  
![image](/uploads/0fe53aa2b6e82367af151c5fecf8279b/image.png){width=80%}

# Solution

The empty footer is an `<nx-section>` that contains/conditionally renders `Account Settings`, `Change Password`, and `Delete User` buttons. The `<nx-section>` also contains an ngIf directive that determines whether it is conditionally rendered. In the particular case of viewing myself as a Cloud User in Webadmin, the conditional logic was such that the `<nx-section>` was being shown:

```
canBeEdited$$() = false
isMe$$() = true
```

when none of the buttons in the section were:

```
!environment.isLocal = false
editPermissions$$().changePassword = false
editPermissions$$().delete = false
```

which is why it appeared to be an empty footer.

This MR updates the conditional rendering logic for the `<nx-section>` to match the rendering logic for each button, so that it only shows when one or more buttons are visible.  
![image](/uploads/e9648c178c34e8017b7c349d8f6fc3fb/image.png){width=80%}