CLOUD-11694  - Clicking the dropdown closes the dropdown if open
## Problem

In the new dropdown, if it is already open and you click the dropdown, then it will stay open.

## Solution

What is really happening is that it is closing and opening the dropdown instantly. You click outside of the dropdown options which closes it, but then you also clicked on the dropdown so it opens again.

To fix this we just need to stop propagation when we handle the event from the overlay.