LIC-292: effectiveState not updating based on ancestor status
# Changes
- added field `effective_state` for `ChannelPartner`, `Organization`, `CloudSystemId`
- added method to change `effective_state` on all descendants in bulk requests
- _django magic_, bulk queries do not call `Model.save()`, so any trigger used in save won't be executed. it allows avoid a lot of performance overhead but requires for control that actions from these triggers are executed on all changed data anyway.
- added triggers on state changes
- services negation refactored to use subqueries instead of `.calculate_current_service()`
- added tests for `state`/`effective_state` changes and update
# Reason for the change
`effective_state` is used to determine a real state of an instate. It depends on states of all ancestors. When it is used dynamically, all parents above are called recursively what is very inefficient. States changes are made very rarely comparing with states invocation so it is more efficient to save `effective_state` to db even if it requires to update all stuff below changed instance.