Remove unused routes in webadmin
## Problem

We have duplicate routes in the WebAdmin specific `NxSystemModule` and also in the Cloud `NxSettingsModule`. The other problem is that `NxSystemModule` imports `NxSettingsModule` AFTER it loads its own routes. Therefor these routes are never even used. The only thing we need to watch out for is Cloud and Webadmin may have specific routes that you can only visit on their respective platforms. We can solve this by using a route guard

## Solution

For the 1 route that is cloud only, we can use a route guard. Otherwise just delete all the old routes as they weren't being used anyway.