CLOUD-11276: Updated sorting for systems
# Problem
Sorting on cloud is not consistent with the other clients.
In addition to this how systems are sorted on cloud portal is not very clear to anyone.

# Solution
Synchronize how sorting is done with the desktop client.
Now we'll prioritize showing online systems and then sort by system name.

## Change
usageFrequency, ownership, ownerName -> status, name