preparing to rebase
Python conversion of ipvd test # 6

Refactored a couple of  ipvd.page functions to return true/false in order to get asserts out of the functions and into the tests. 

Moved some of the functions that operated on the ipvd table into the IPVDTable object.

Refactored validate_device_table_not_empty() function so that code to count row is separate.validate_device_table_not_empty() was returning the number of rows, even though that was not always used.