CLOUD-11637: Update resource permissions checks
# Problem
Our frontend user permissions are currently implemented such that a user is given the aggregate of the permissions across all the individual resources and resource groups they have access to. This has caused some issues, for instance for system health viewers being able to see the "View" and "Layouts" tab, since they have view access rights for a server (not a device).

Another issue was that we were storing the aggregate permissions as a pipe-separated string (eg `view|viewBookmarks|editArchives`) and then using a `.includes()` check on that string. This caused a bug where a user with `viewBookmarks` permissions was interpreted as also having `view` permissions, since `viewBookmarks.includes("view") === true`.

# Solution
This MR includes some updates to how we are compiling/storing the user's permissions, to allow more granular resource-based checks across the codebase, and to address the issues above:

1. add a `resourceAccessRights$$` signal to store resource-related permissions as an object with booleans rather than a pipe-separated string
2. when calculating any resource-related permission in the `permissions$$` signal, use the user's permissions from just the device group, rather than aggregating all their resource permissions
3. create helper methods for checking permissions for a specific device. These methods check whether the user has permissions from either their groups (`permissions$$`) or resource-specific/custom permissions (`resourceAccessRights$$`)
4. add a `canViewADevice` helper method on the system module, which checks if a user is able to view one or more devices in the system (either due to group permissions or resource-specific permissions).

## Fixes:
- CLOUD-11330
- CLOUD-11432
- CLOUD-11592
- CLOUD-11685
- CLOUD-11625
- CLOUD-11471