Process button replacement
## Changes
A replacement for the old process button(s) in dialogs. Same behavior, where it goes into a loading state and performs an async action, with success and failure callbacks to handle the result. Only for dialogs at the moment, will probably make it usable elsewhere later.

Differences to the old button:
- Fully typed, so no more [magical behind the scenes error handling](https://gitlab.nxvms.dev/dev/cloud_portal/-/blob/f0ba28343bde5fda72185e4965f86562f323b575/front_end/libs/services/process.service/process.ts#L32)
- No "cancelling" a request. The previous behavior stopped the front end from listening for a response, but didn't actually stop the server from recieving and acting on the request
- Uses `<ng-content>` instead of inputs so no [hardcoded content styling](https://gitlab.nxvms.dev/dev/cloud_portal/-/blob/develop/front_end/libs/components/process-button/process-button.component.html#L35)
- Built-in syncing of busy state which can handle locking/unlocking dialogs automatically

## TODO
Update button styles to not change on click while disabled